+++ wallet-connect-test-bench/front-end/src/Main.tsx
@@ -51,6 +51,7 @@ import {
   simpleCCDTransfer,
   simpleCCDTransferToNonExistingAccountAddress,
   sponsorSetU8,
+  sponsorConfigureDelegation,
 } from "./writing_to_blockchain";
 
 import {
@@ -1544,6 +1545,65 @@ export default function Main(props: MainProps) {
                     }}
                   >
                     Submit (sponsored transaction to update a smart contract)
+                  </button>
+                </TestBox>
+                <TestBox
+                  header="(TE) Testing that configure delegation can be done as part of a sponsored transaction"
+                  note="
+                                        Expected result after pressing the button and confirming in wallet: The
+                                        transaction hash or an error message should appear in the right column.
+                                        "
+                >
+                                    <label className="field">
+                    <p>Sponsor Account:</p>
+                    <input
+                      className="inputFieldStyle"
+                      id="sponsorAccount"
+                      type="text"
+                      placeholder="Sponsor account address"
+                      value={sponsorAccount}
+                      onChange={(e) => setSponsorAccount(e.target.value)}
+                    />
+                  </label>
+                  <br />
+                  <label className="field">
+                    <p>Sponsor Private Key:</p>
+                    <input
+                      className="inputFieldStyle"
+                      id="sponsorPrivateKey"
+                      type="text"
+                      placeholder="Sponsor private key"
+                      value={sponsorPrivateKey}
+                      onChange={(e) =>
+                        setSponsorPrivateKey(e.target.value)
+                      }
+                    />
+                  </label>
+                  <br />
+                   <label className="field">
+                    CCD stake (micro):
+                    <br />
+                    <input
+                      className="inputFieldStyle"
+                      id="CCDAmount"
+                      type="text"
+                      placeholder="0"
+                      onChange={changeCCDAmountHandler}
+                    />
+                  </label>
+                  <button
+                    className="btn btn-primary"
+                    type="button"
+                    onClick={() => {
+                      setTxHash("");
+                      setTransactionError("");
+                      const tx = sponsorConfigureDelegation(connection, account, sponsorAccount, sponsorPrivateKey, cCDAmount, submitPayloadToSponsor);
+                      tx.then(setTxHash).catch((err: Error) =>
+                        setTransactionError((err as Error).message)
+                      );
+                    }}
+                  >
+                    Submit (sponsored transaction to configure delegation)
                   </button>
                 </TestBox>
                 <br />