+++ wallet-connect-test-bench/front-end/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Add a new section on the UI to sponsor the transaction configuring delegation.
+
 ## 1.5.7 (1.5.6 abandoned)
 
 - Add a new section on the UI to sponsor the transaction invoking a smart contract